-implemented collision between the enemies and projectiles -minor cleanup

+++ src/entity/enemy/special_enemy/dev/DevEnemy.h
@@ -9,7 +9,7 @@ class DevEnemy final : public Enemy{
     public:
         ~DevEnemy() override = default;
 
-        DevEnemy(sf::VertexArray *pathToFollow, const float speed, const int health, const sf::Color color, const int size) : Enemy(pathToFollow, speed, health) {
+        DevEnemy(sf::VertexArray *pathToFollow, const float speed, const int health, const sf::Color color, const float size) : Enemy(pathToFollow, speed, health) {
             getHitTexture()->getDisplayEntity()->setFillColor(color);
             getHitTexture()->getDisplayEntity()->setRadius(size);
             getHitTexture()->getHitbox()->setFillColor(sf::Color::Transparent);
